+++ packages/gcf-utils/src/gcf-utils.ts
@@ -18,6 +18,7 @@ import {createProbotAuth} from 'octokit-auth-probot';
 
 import getStream from 'get-stream';
 import intoStream from 'into-stream';
+import * as http from 'http';
 
 import {v1 as SecretManagerV1} from '@google-cloud/secret-manager';
 import {v2 as CloudTasksV2} from '@google-cloud/tasks';
@@ -29,9 +30,15 @@ import {config as ConfigPlugin} from '@probot/octokit-plugin-config';
 import {buildTriggerInfo} from './logging/trigger-info-builder';
 import {GCFLogger} from './logging/gcf-logger';
 import {v4} from 'uuid';
+import {getServer} from './server/server';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const LoggingOctokitPlugin = require('../src/logging/logging-octokit-plugin.js');
+
+export type HandlerFunction = (
+  request: express.Request,
+  response: express.Response
+) => Promise<void>;
 
 type CronType = 'repository' | 'installation' | 'global';
 const DEFAULT_CRON_TYPE: CronType = 'repository';
@@ -64,16 +71,42 @@ interface EnqueueTaskParams {
 export interface WrapOptions {
   // Whether or not to enqueue direct GitHub webhooks in a Cloud Task
   // queue which provides a retry mechanism. Defaults to `true`.
-  background: boolean;
+  // Deprecated. Please use `maxRetries` and `maxCronRetries` instead.
+  background?: boolean;
 
   // Whether or not to automatically log Octokit requests. Defaults to
   // `false`.
-  logging: boolean;
+  logging?: boolean;
 
   // Whether or not to skip verification of request payloads. Defaults
   // to `true` in test mode, otherwise `false`.
   skipVerification?: boolean;
+
+  // Maximum number of attempts for webhook handlers. Defaults to `10`.
+  maxRetries?: number;
+
+  // Maximum number of attempts for cron handlers. Defaults to `0`.
+  maxCronRetries?: number;
+
+  // Maximum number of attempts for pubsub handlers. Defaults to `0`.
+  maxPubSubRetries?: number;
 }
+
+interface WrapConfig {
+  logging: boolean;
+  skipVerification: boolean;
+  maxCronRetries: number;
+  maxRetries: number;
+  maxPubSubRetries: number;
+}
+
+const DEFAULT_WRAP_CONFIG: WrapConfig = {
+  logging: false,
+  skipVerification: RUNNING_IN_TEST,
+  maxCronRetries: 0,
+  maxRetries: 10,
+  maxPubSubRetries: 0,
+};
 
 export const logger = new GCFLogger();
 
@@ -163,18 +196,62 @@ export const addOrUpdateIssueComment = async (
   }
 };
 
+interface BootstrapperOptions {
+  secretsClient?: SecretManagerV1.SecretManagerServiceClient;
+  tasksClient?: CloudTasksV2.CloudTasksClient;
+  projectId?: string;
+  functionName?: string;
+  location?: string;
+  payloadBucket?: string;
+}
+
 export class GCFBootstrapper {
   probot?: Probot;
 
   secretsClient: SecretManagerV1.SecretManagerServiceClient;
   cloudTasksClient: CloudTasksV2.CloudTasksClient;
   storage: Storage;
+  projectId: string;
+  functionName: string;
+  location: string;
+  payloadBucket: string | undefined;
 
-  constructor(secretsClient?: SecretManagerV1.SecretManagerServiceClient) {
+  constructor(options?: BootstrapperOptions) {
+    options = {
+      ...{
+        projectId: process.env.PROJECT_ID,
+        functionName: process.env.GCF_SHORT_FUNCTION_NAME,
+        location: process.env.GCF_LOCATION,
+        payloadBucket: process.env.WEBHOOK_TMP,
+      },
+      ...options,
+    };
+
     this.secretsClient =
-      secretsClient || new SecretManagerV1.SecretManagerServiceClient();
-    this.cloudTasksClient = new CloudTasksV2.CloudTasksClient();
+      options?.secretsClient ||
+      new SecretManagerV1.SecretManagerServiceClient();
+    this.cloudTasksClient =
+      options?.tasksClient || new CloudTasksV2.CloudTasksClient();
     this.storage = new Storage({autoRetry: !RUNNING_IN_TEST});
+    if (!options.projectId) {
+      throw new Error(
+        'Missing required `projectId`. Please provide as a constructor argument or set the PROJECT_ID env variable.'
+      );
+    }
+    this.projectId = options.projectId;
+    if (!options.functionName) {
+      throw new Error(
+        'Missing required `functionName`. Please provide as a constructor argument or set the GCF_SHORT_FUNCTION_NAME env variable.'
+      );
+    }
+    this.functionName = options.functionName;
+    if (!options.location) {
+      throw new Error(
+        'Missing required `location`. Please provide as a constructor argument or set the GCF_LOCATION env variable.'
+      );
+    }
+    this.location = options.location;
+    this.payloadBucket = options.payloadBucket;
   }
 
   async loadProbot(
@@ -192,9 +269,7 @@ export class GCFBootstrapper {
   }
 
   getSecretName(): string {
-    const projectId = process.env.PROJECT_ID || '';
-    const functionName = process.env.GCF_SHORT_FUNCTION_NAME || '';
-    return `projects/${projectId}/secrets/${functionName}`;
+    return `projects/${this.projectId}/secrets/${this.functionName}`;
   }
 
   getLatestSecretVersionName(): string {
@@ -269,6 +344,7 @@ export class GCFBootstrapper {
     id: string;
     signature: string;
     taskId: string;
+    taskRetries: number;
   } {
     const name =
       request.get('x-github-event') || request.get('X-GitHub-Event') || '';
@@ -281,7 +357,12 @@ export class GCFBootstrapper {
       request.get('X-CloudTasks-TaskName') ||
       request.get('x-cloudtasks-taskname') ||
       '';
-    return {name, id, signature, taskId};
+    const taskRetries = parseInt(
+      request.get('X-CloudTasks-TaskRetryCount') ||
+        request.get('x-cloudtasks-taskretrycount') ||
+        '0'
+    );
+    return {name, id, signature, taskId, taskRetries};
   }
 
   /**
@@ -302,23 +383,59 @@ export class GCFBootstrapper {
     return TriggerType.UNKNOWN;
   }
 
-  gcf(
-    appFn: ApplicationFunction,
-    wrapOptions?: WrapOptions
-  ): (request: express.Request, response: express.Response) => Promise<void> {
-    return async (request: express.Request, response: express.Response) => {
-      wrapOptions = wrapOptions ?? {background: true, logging: false};
+  private parseWrapConfig(wrapOptions: WrapOptions | undefined): WrapConfig {
+    const wrapConfig: WrapConfig = {
+      ...DEFAULT_WRAP_CONFIG,
+      ...wrapOptions,
+    };
 
-      // default skip verification for testing
-      // to test signatures, explicitly set skipVerification to false
-      if (wrapOptions.skipVerification === undefined) {
-        wrapOptions.skipVerification = RUNNING_IN_TEST;
+    if (wrapOptions?.background !== undefined) {
+      logger.warn(
+        '`background` option has been deprecated in favor of `maxRetries` and `maxCronRetries`'
+      );
+      if (wrapOptions.background === false) {
+        wrapConfig.maxCronRetries = 0;
+        wrapConfig.maxRetries = 0;
+        wrapConfig.maxPubSubRetries = 0;
       }
+    }
+    return wrapConfig;
+  }
+
+  private getRetryLimit(wrapConfig: WrapConfig, eventName: string) {
+    if (eventName.startsWith('schedule.')) {
+      return wrapConfig.maxCronRetries;
+    }
+    if (eventName.startsWith('pubsub.')) {
+      return wrapConfig.maxPubSubRetries;
+    }
+    return wrapConfig.maxRetries;
+  }
+
+  /**
+   * Wrap an ApplicationFunction in a http.Server that can be started
+   * directly.
+   * @param appFn {ApplicationFunction} The probot handler function
+   * @param wrapOptions {WrapOptions} Bot handler options
+   */
+  server(appFn: ApplicationFunction, wrapOptions?: WrapOptions): http.Server {
+    return getServer(this.gcf(appFn, wrapOptions));
+  }
+
+  /**
+   * Wrap an ApplicationFunction in so it can be started in a Google
+   * Cloud Function.
+   * @param appFn {ApplicationFunction} The probot handler function
+   * @param wrapOptions {WrapOptions} Bot handler options
+   */
+  gcf(appFn: ApplicationFunction, wrapOptions?: WrapOptions): HandlerFunction {
+    return async (request: express.Request, response: express.Response) => {
+      const wrapConfig = this.parseWrapConfig(wrapOptions);
 
       this.probot =
-        this.probot || (await this.loadProbot(appFn, wrapOptions.logging));
+        this.probot || (await this.loadProbot(appFn, wrapConfig.logging));
 
-      const {name, id, signature, taskId} =
+      const {name, id, signature, taskId, taskRetries} =
         GCFBootstrapper.parseRequestHeaders(request);
 
       const triggerType: TriggerType = GCFBootstrapper.parseTriggerType(
@@ -328,7 +445,7 @@ export class GCFBootstrapper {
 
       // validate the signature
       if (
-        !wrapOptions.skipVerification &&
+        !wrapConfig.skipVerification &&
         !this.probot.webhooks.verify(request.body, signature)
       ) {
         response.send({
@@ -350,28 +467,35 @@ export class GCFBootstrapper {
           return;
         } else if (triggerType === TriggerType.SCHEDULER) {
           // Cloud scheduler tasks (cron)
-          await this.handleScheduled(id, request, wrapOptions);
-        } else if (
-          triggerType === TriggerType.TASK ||
-          triggerType === TriggerType.PUBSUB ||
-          !wrapOptions?.background
-        ) {
-          if (!wrapOptions?.background) {
-            // a bot can opt out of running through tasks, some bots do this
-            // due to large payload sizes:
-            logger.info(`${id}: skipping Cloud Tasks`);
+          await this.handleScheduled(id, request, wrapConfig);
+        } else if (triggerType === TriggerType.PUBSUB) {
+          const payload = this.parsePubSubPayload(request);
+          await this.enqueueTask({
+            id,
+            name,
+            body: JSON.stringify(payload),
+          });
+        } else if (triggerType === TriggerType.TASK) {
+          const maxRetries = this.getRetryLimit(wrapConfig, name);
+          // Abort task retries if we've hit the max number by
+          // returning "success"
+          if (taskRetries > maxRetries) {
+            logger.metric('too-many-retries');
+            logger.info(`Too many retries: ${taskRetries} > ${maxRetries}`);
+            response.send({
+              statusCode: 200,
+              body: JSON.stringify({message: 'Too many retries'}),
+            });
+            return;
           }
-          let payload = request.body;
-          if (triggerType === TriggerType.PUBSUB) {
-            payload = this.parsePubSubPayload(request);
-          }
+
           // If the payload contains `tmpUrl` this indicates that the original
           // payload has been written to Cloud Storage; download it.
-          const body = await this.maybeDownloadOriginalBody(payload);
+          const payload = await this.maybeDownloadOriginalBody(request.body);
 
           // The payload does not exist, stop retrying on this task by letting
           // this request "succeed".
-          if (!body) {
+          if (!payload) {
             logger.metric('payload-expired');
             response.send({
               statusCode: 200,
@@ -386,7 +510,7 @@ export class GCFBootstrapper {
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
             name: name as any,
             id,
-            payload: body,
+            payload,
           });
         } else if (triggerType === TriggerType.GITHUB) {
           await this.enqueueTask({
@@ -420,21 +544,21 @@ export class GCFBootstrapper {
    * @param body {Scheduled} Scheduler params. May contain additional request
    *   parameters besides the ones defined by the Scheduled type.
    * @param signature
-   * @param wrapOptions
+   * @param wrapConfig
    */
   private async handleScheduled(
     id: string,
     req: express.Request,
-    wrapOptions: WrapOptions
+    wrapConfig: WrapConfig
   ) {
     const body: Scheduled = this.parseRequestBody(req);
     const cronType = body.cron_type ?? DEFAULT_CRON_TYPE;
     if (cronType === 'global') {
       await this.handleScheduledGlobal(id, body);
     } else if (cronType === 'installation') {
-      await this.handleScheduledInstallation(id, body, wrapOptions);
+      await this.handleScheduledInstallation(id, body, wrapConfig);
     } else {
-      await this.handleScheduledRepository(id, body, wrapOptions);
+      await this.handleScheduledRepository(id, body, wrapConfig);
     }
   }
 
@@ -459,15 +583,22 @@ export class GCFBootstrapper {
    * Async iterator over each installation for an app.
    *
    * See https://docs.github.com/en/rest/reference/apps#list-installations-for-the-authenticated-app
-   * @param wrapOptions {WrapOptions}
+   * @param wrapConfig {WrapConfig}
    */
-  private async *eachInstallation(wrapOptions: WrapOptions) {
-    const octokit = await this.getAuthenticatedOctokit(undefined, wrapOptions);
+  private async *eachInstallation(wrapConfig: WrapConfig) {
+    const octokit = await this.getAuthenticatedOctokit(undefined, wrapConfig);
     const installationsPaginated = octokit.paginate.iterator(
       octokit.apps.listInstallations
     );
     for await (const response of installationsPaginated) {
       for (const installation of response.data) {
+        if (installation.suspended_at !== null) {
+          // Assume the installation is suspended.
+          logger.info(
+            `skipping installations for ${installation.id} because it is suspended`
+          );
+          continue;
+        }
         yield installation;
       }
     }
@@ -477,15 +608,15 @@ export class GCFBootstrapper {
    * Async iterator over each repository for an app installation.
    *
    * See https://docs.github.com/en/rest/reference/apps#list-repositories-accessible-to-the-app-installation
-   * @param wrapOptions {WrapOptions}
+   * @param wrapConfig {WrapConfig}
    */
   private async *eachInstalledRepository(
     installationId: number,
-    wrapOptions: WrapOptions
+    wrapConfig: WrapConfig
   ) {
     const octokit = await this.getAuthenticatedOctokit(
       installationId,
-      wrapOptions
+      wrapConfig
     );
     const installationRepositoriesPaginated = octokit.paginate.iterator(
       octokit.apps.listReposAccessibleToInstallation,
@@ -513,12 +644,12 @@ export class GCFBootstrapper {
    * @param id {string} GitHub delivery GUID
    * @param body {Scheduled} Scheduler params. May contain additional request
    *   parameters besides the ones defined by the Scheduled type.
-   * @param wrapOptions
+   * @param wrapConfig
    */
   private async handleScheduledInstallation(
     id: string,
     body: Scheduled,
-    wrapOptions: WrapOptions
+    wrapConfig: WrapConfig
   ) {
     if (body.installation) {
       await this.enqueueTask({
@@ -527,7 +658,7 @@ export class GCFBootstrapper {
         body: JSON.stringify(body),
       });
     } else {
-      const generator = this.eachInstallation(wrapOptions);
+      const generator = this.eachInstallation(wrapConfig);
       for await (const installation of generator) {
         const extraParams: Scheduled = {
           installation: {
@@ -567,12 +698,12 @@ export class GCFBootstrapper {
    * @param body {Scheduled} Scheduler params. May contain additional request
    *   parameters besides the ones defined by the Scheduled type.
    * @param signature
-   * @param wrapOptions
+   * @param wrapConfig
    */
   private async handleScheduledRepository(
     id: string,
     body: Scheduled,
-    wrapOptions: WrapOptions
+    wrapConfig: WrapConfig
   ) {
     if (body.repo) {
       // Job was scheduled for a single repository:
@@ -585,27 +716,40 @@ export class GCFBootstrapper {
     } else if (body.installation) {
       const generator = this.eachInstalledRepository(
         body.installation.id,
-        wrapOptions
+        wrapConfig
       );
+      const promises: Array<Promise<void>> = new Array<Promise<void>>();
+      const batchSize = 30;
       for await (const repo of generator) {
         if (repo.archived === true || repo.disabled === true) {
           continue;
         }
-        await this.scheduledToTask(
-          repo.full_name,
-          id,
-          body,
-          SCHEDULER_REPOSITORY_EVENT_NAME
+        promises.push(
+          this.scheduledToTask(
+            repo.full_name,
+            id,
+            body,
+            SCHEDULER_REPOSITORY_EVENT_NAME
+          )
         );
+        if (promises.length >= batchSize) {
+          await Promise.all(promises);
+          promises.splice(0, promises.length);
+        }
+      }
+      // Wait for the rest.
+      if (promises.length > 0) {
+        await Promise.all(promises);
+        promises.splice(0, promises.length);
       }
     } else {
-      const installationGenerator = this.eachInstallation(wrapOptions);
+      const installationGenerator = this.eachInstallation(wrapConfig);
       const promises: Array<Promise<void>> = new Array<Promise<void>>();
-      const batchNum = 30;
+      const batchSize = 30;
       for await (const installation of installationGenerator) {
         const generator = this.eachInstalledRepository(
           installation.id,
-          wrapOptions
+          wrapConfig
         );
         const extraParams: Scheduled = {
           installation: {
@@ -635,7 +779,7 @@ export class GCFBootstrapper {
               SCHEDULER_REPOSITORY_EVENT_NAME
             )
           );
-          if (promises.length >= batchNum) {
+          if (promises.length >= batchSize) {
             await Promise.all(promises);
             promises.splice(0, promises.length);
           }
@@ -655,19 +799,24 @@ export class GCFBootstrapper {
    * @param installationId {number|undefined} The installation id to
    *   authenticate as. Required if you are trying to take action
    *   on an installed repository.
-   * @param wrapOptions
+   * @param wrapConfig
    */
   async getAuthenticatedOctokit(
     installationId?: number,
-    wrapOptions?: WrapOptions
+    wrapConfig?: WrapConfig
   ): Promise<Octokit> {
-    const cfg = await this.getProbotConfig(wrapOptions?.logging);
-    const opts = {
+    const cfg = await this.getProbotConfig(wrapConfig?.logging);
+    let opts = {
       appId: cfg.appId,
       privateKey: cfg.privateKey,
-      installationId,
     };
-    if (wrapOptions?.logging) {
+    if (installationId) {
+      opts = {
+        ...opts,
+        ...{installationId},
+      };
+    }
+    if (wrapConfig?.logging) {
       const LoggingOctokit = Octokit.plugin(LoggingOctokitPlugin)
         .plugin(ConfigPlugin)
         .defaults({authStrategy: createProbotAuth});
@@ -750,20 +899,15 @@ export class GCFBootstrapper {
   async enqueueTask(params: EnqueueTaskParams) {
     logger.info('scheduling cloud task');
     // Make a task here and return 200 as this is coming from GitHub
-    const projectId = process.env.PROJECT_ID || '';
-    const location = process.env.GCF_LOCATION || '';
     // queue name can contain only letters ([A-Za-z]), numbers ([0-9]), or hyphens (-):
-    const queueName = (process.env.GCF_SHORT_FUNCTION_NAME || '').replace(
-      /_/g,
-      '-'
-    );
+    const queueName = this.functionName.replace(/_/g, '-');
     const queuePath = this.cloudTasksClient.queuePath(
-      projectId,
-      location,
+      this.projectId,
+      this.location,
       queueName
     );
     // https://us-central1-repo-automation-bots.cloudfunctions.net/merge_on_green:
-    const url = `https://${location}-${projectId}.cloudfunctions.net/${process.env.GCF_SHORT_FUNCTION_NAME}`;
+    const url = `https://${this.location}-${this.projectId}.cloudfunctions.net/${this.functionName}`;
     logger.info(`scheduling task in queue ${queueName}`);
     if (params.body) {
       // Payload conists of either the original params.body or, if Cloud
@@ -814,9 +958,9 @@ export class GCFBootstrapper {
    * @param body
    */
   private async maybeWriteBodyToTmp(body: string): Promise<string> {
-    if (process.env.WEBHOOK_TMP) {
+    if (this.payloadBucket) {
       const tmp = `${Date.now()}-${v4()}.txt`;
-      const bucket = this.storage.bucket(process.env.WEBHOOK_TMP);
+      const bucket = this.storage.bucket(this.payloadBucket);
       const writeable = bucket.file(tmp).createWriteStream({
         validation: !RUNNING_IN_TEST,
       });
@@ -844,13 +988,13 @@ export class GCFBootstrapper {
     [key: string]: string;
   }): Promise<object | null> {
     if (payload.tmpUrl) {
-      if (!process.env.WEBHOOK_TMP) {
+      if (!this.payloadBucket) {
         throw Error('no tmp directory configured');
       }
-      const bucket = this.storage.bucket(process.env.WEBHOOK_TMP);
+      const bucket = this.storage.bucket(this.payloadBucket);
       const file = bucket.file(payload.tmpUrl);
       const readable = file.createReadStream({
-        validation: process.env.NODE_ENV !== 'test',
+        validation: !RUNNING_IN_TEST,
       });
       try {
         const content = await getStream(readable);