+++ packages/sync-repo-settings/test/test.bot.ts
@@ -176,6 +176,7 @@ describe('Sync repo settings', () => {
           'test (15)',
           'cla/google',
           'windows',
+          'OwlBot Post Processor',
         ],
         true,
         true